Move RuleViolation to sorald.rule package
