+++ src/main/java/sorald/Repair.java
@@ -24,11 +24,11 @@ import sorald.event.SoraldEventHandler;
 import sorald.event.collectors.CompilationUnitCollector;
 import sorald.event.models.CrashEvent;
 import sorald.processor.SoraldAbstractProcessor;
+import sorald.rule.RuleViolation;
 import sorald.segment.FirstFitSegmentationAlgorithm;
 import sorald.segment.Node;
 import sorald.segment.SoraldTreeBuilderAlgorithm;
 import sorald.sonar.BestFitScanner;
-import sorald.sonar.RuleViolation;
 import spoon.Launcher;
 import spoon.MavenLauncher;
 import spoon.compiler.Environment;
