+++ src/main/java/sorald/event/models/RepairEvent.java
@@ -2,7 +2,7 @@ package sorald.event.models;
 
 import sorald.event.EventType;
 import sorald.event.SoraldEvent;
-import sorald.sonar.RuleViolation;
+import sorald.rule.RuleViolation;
 import spoon.reflect.declaration.CtElement;
 
 /**
