+++ src/test/java/sorald/processor/BestFitScannerTest.java
@@ -13,10 +13,10 @@ import java.util.Set;
 import org.junit.jupiter.api.Test;
 import sorald.Constants;
 import sorald.TestHelper;
+import sorald.rule.RuleViolation;
 import sorald.sonar.BestFitScanner;
 import sorald.sonar.Checks;
 import sorald.sonar.ProjectScanner;
-import sorald.sonar.RuleViolation;
 import spoon.FluentLauncher;
 import spoon.Launcher;
 import spoon.reflect.CtModel;
